refactor: take out slider on item page in own component
